requirements.txt
aiosmtplib==3.0.2
alembic==1.14.0
annotated-types==0.7.0
anyio==4.8.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
Authlib==1.5.2
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
cssselect==1.2.0
cssutils==2.11.1
Deprecated==1.2.18
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.6
fastapi-cli==0.0.7
fastapi-mail==1.4.2
gcloud==0.18.3
googleapis-common-protos==1.70.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jwcrypto==1.5.6
limits==5.4.0
lxml==5.3.0
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
minio==7.2.15
more-itertools==10.5.0
oauth2client==4.1.3
orjson==3.10.14
packaging==25.0
passlib==1.7.4
pdfkit==1.0.0
premailer==3.10.0
protobuf==6.30.2
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.10.5
pydantic-extra-types==2.10.1
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.3
Pyrebase4==4.8.0
python-decouple==3.8
python-dotenv==1.0.1
python-jose==3.3.0
python-jwt==4.1.0
python-multipart==0.0.20
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==0.10.1
rich==13.9.4
rich-toolkit==0.12.0
rsa==4.9
shellingham==1.5.4
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.41.3
text-unidecode==1.3
typer==0.15.1
typing_extensions==4.12.2
ujson==5.10.0
urllib3==1.26.20
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.3
websockets==14.1
wrapt==1.17.2
